Finished most of resume.
Finished most of my resume, adding in some new sections and creating/
updating Liquid templates to reflect changes. Added a new Liquid tag
for "buzzword capsules" which will need to be styled in the next commit
and eventually support some ambitious ideas I want to do with a very
JS-heavy resume page.

Added a logo to the site in various sizes so that a responsive image
solution can be implemented. Added logo in config file, which should be
finished at this point.

Added RSpec to the solution because I foresee some logic requiring tests
in the near-future as I work on the JS resume tweaks.

Added MIT license to repository and created a credits.md file to track
open source libraries/modules/whatever I am using and linking to the
original repository or the contributor's site along with displaying a
copy of their licensing terms.

Signed-off-by: Scott Simontis <yo@scottsimontis.io>